+++ CHANGELOG.md
@@ -36,7 +36,7 @@ This changelog keeps track of work items that have been completed and are ready 
 
 ### Fixes
 
-- **General**: TODO ([#TODO](https://github.com/kedacore/http-add-on/issues/TODO))
+- **Interceptor**: Preserve X-Forwarded-Proto from upstream ([#1432](https://github.com/kedacore/http-add-on/issues/1432))
 
 ### Deprecations